+++ src/structural_typing/assist/predicate_defining.clj
@@ -21,3 +21,21 @@
    (format "(exactly %s)" (readable/value-string expected))
    (partial = expected)
    (should-be "%s should be exactly `%s`; it is `%s`" expected)))
+
+(defn regex-match [expected]
+  (compose-predicate
+   (format "(regex-match %s)" (str expected))
+   (fn [actual]
+     (branch-on actual
+       string?   (boolean (re-find expected actual))
+       regex?    (= (str expected) (str actual))
+       :else     false))
+   (fn [oopsie]
+     (let [actual (:leaf-value oopsie)
+           readable-actual (readable/value-string actual)]
+       (format "%s should match %s; it is %s"
+               (oopsie/friendly-path oopsie)
+               (readable/value-string expected)
+               (if (or (string? actual) (regex? actual))
+                 readable-actual
+                 (str "`" readable-actual "`")))))))