+++ CHANGELOG.md
@@ -5,7 +5,8 @@ See [here](http://keepachangelog.com/) for the change log format.
 ## [in progress]
 - ADD: Strings can now be used in paths. Like keywords, they denote map keywords to follow.
 - ADD: `ONLY` is like `ALL`, but it insists that its collection have exactly one element.
-- ADD: A non-function `x` used in predicate position behaves as `(exactly x)`. [intermediate step]
+- ADD: By default, a non-function `x` used in predicate position behaves as `(exactly x)`.
+       - Regular expressions are treated as in Midje.
 - CHANGE: Some improvements to error handling
 
 ## [1.0.1]